Parse uploaded files with unset MIME types.
